Enable `eslint-plugin-custom-elements`
