+++ package.json
@@ -26,6 +26,7 @@
     "@github/prettier-config": "0.0.4",
     "chai": "^4.3.4",
     "eslint": "^7.26.0",
+    "eslint-plugin-custom-elements": "^0.0.2",
     "eslint-plugin-github": "^4.1.3",
     "karma": "^6.3.2",
     "karma-chai": "^0.1.0",
