+++ tsconfig.json
@@ -7,7 +7,5 @@
     "outDir": "dist",
     "removeComments": true
   },
-  "files": [
-    "src/index.ts"
-  ]
+  "files": ["src/file-attachment-element.ts", "src/attachment.ts"]
 }
